folder + dictionary
